[refactor] refactor CPU to only use device functions so there is no more hard coded logic in transformer logic.
Also refactored view/mutview file to provide proper abstraction for both CPU
and GPU. A key change is to pass down impl to the corresponding than
handling in device.

For example: instead returning &[f32] from slice() or as_ref() function,
we only store range & data reference in View and it is the underlying
device's responsibilities to extract out the correct portion of the data
for calculation.